refactor(internal): Rename summary to stitch
Rename the 'summary' package to 'stitch',
since it'll hold other core data types for stitchmd.
