+++ internal/stitch/testdata/parse_errors.yaml
